mudando o tipo de dado de int para float na calculadora

+++ week-0/calculator.py
@@ -1,6 +1,6 @@
 # Variables to start the calculator
-x = int(input("What's x? "))
-y = int(input("What's y? "))
+x = float(input("What's x? "))
+y = float(input("What's y? "))
 
 # Printing the results
 print(x + y)
